requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
jinja2==3.1.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.3.0
pytest==7.4.2
pytest-asyncio==0.22.0
httpx==0.28
alembic==1.11.1
aiofiles==23.2.1
python-multipart==0.0.6
email-validator==2.2.0